# docs/openapi.json
{
  "components": {
    "schemas": {
      "Body_create_import_v1_imports_post": {
        "properties": {
          "file": {
            "contentMediaType": "application/octet-stream",
            "description": "UTF-8 customer CSV",
            "title": "File",
            "type": "string"
          }
        },
        "required": [
          "file"
        ],
        "title": "Body_create_import_v1_imports_post",
        "type": "object"
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
      "IntegrationResponse": {
        "properties": {
          "attempt_count": {
            "title": "Attempt Count",
            "type": "integer"
          },
          "available_at": {
            "format": "date-time",
            "title": "Available At",
            "type": "string"
          },
          "delivered_at": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Delivered At"
          },
          "event_id": {
            "format": "uuid",
            "title": "Event Id",
            "type": "string"
          },
          "event_type": {
            "title": "Event Type",
            "type": "string"
          },
          "last_failure_code": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Last Failure Code"
          },
          "operation_id": {
            "format": "uuid",
            "title": "Operation Id",
            "type": "string"
          },
          "status": {
            "title": "Status",
            "type": "string"
          }
        },
        "required": [
          "event_id",
          "operation_id",
          "event_type",
          "status",
          "attempt_count",
          "available_at",
          "last_failure_code",
          "delivered_at"
        ],
        "title": "IntegrationResponse",
        "type": "object"
      },
      "IssueResponse": {
        "properties": {
          "code": {
            "title": "Code",
            "type": "string"
          },
          "correction": {
            "title": "Correction",
            "type": "string"
          },
          "field": {
            "title": "Field",
            "type": "string"
          },
          "message": {
            "title": "Message",
            "type": "string"
          },
          "row": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Row"
          }
        },
        "required": [
          "row",
          "field",
          "code",
          "message",
          "correction"
        ],
        "title": "IssueResponse",
        "type": "object"
      },
      "OperationResponse": {
        "properties": {
          "accepted": {
            "title": "Accepted",
            "type": "boolean"
          },
          "attempt_count": {
            "title": "Attempt Count",
            "type": "integer"
          },
          "duration_ms": {
            "title": "Duration Ms",
            "type": "integer"
          },
          "error_codes": {
            "items": {
              "type": "string"
            },
            "title": "Error Codes",
            "type": "array"
          },
          "existing_rows": {
            "title": "Existing Rows",
            "type": "integer"
          },
          "inserted_rows": {
            "title": "Inserted Rows",
            "type": "integer"
          },
          "issues": {
            "items": {
              "$ref": "#/components/schemas/IssueResponse"
            },
            "title": "Issues",
            "type": "array"
          },
          "operation_id": {
            "format": "uuid",
            "title": "Operation Id",
            "type": "string"
          },
          "status": {
            "title": "Status",
            "type": "string"
          },
          "total_rows": {
            "title": "Total Rows",
            "type": "integer"
          }
        },
        "required": [
          "operation_id",
          "accepted",
          "status",
          "total_rows",
          "inserted_rows",
          "existing_rows",
          "duration_ms",
          "attempt_count",
          "error_codes",
          "issues"
        ],
        "title": "OperationResponse",
        "type": "object"
      },
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      }
    },
    "securitySchemes": {
      "BearerAuth": {
        "scheme": "bearer",
        "type": "http"
      }
    }
  },
  "info": {
    "description": "Strict imports with durable asynchronous partner notifications.",
    "title": "Customer Import API",
    "version": "0.3.1"
  },
  "openapi": "3.1.0",
  "paths": {
    "/health/live": {
      "get": {
        "operationId": "live_health_live_get",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": {
                    "type": "string"
                  },
                  "title": "Response Live Health Live Get",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          }
        },
        "summary": "Live",
        "tags": [
          "health"
        ]
      }
    },
    "/health/ready": {
      "get": {
        "operationId": "ready_health_ready_get",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": {
                    "type": "string"
                  },
                  "title": "Response Ready Health Ready Get",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          }
        },
        "summary": "Ready",
        "tags": [
          "health"
        ]
      }
    },
    "/v1/imports": {
      "post": {
        "operationId": "create_import_v1_imports_post",
        "parameters": [
          {
            "in": "header",
            "name": "Idempotency-Key",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Idempotency-Key"
            }
          }
        ],
        "requestBody": {
          "content": {
            "multipart/form-data": {
              "schema": {
                "$ref": "#/components/schemas/Body_create_import_v1_imports_post"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OperationResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "202": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OperationResponse"
                }
              }
            },
            "description": "Identical request is processing."
          },
          "400": {
            "description": "Bad Request"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "409": {
            "description": "Conflict"
          },
          "413": {
            "description": "Request Entity Too Large"
          },
          "422": {
            "description": "Unprocessable Entity"
          },
          "503": {
            "description": "Service Unavailable"
          }
        },
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "summary": "Create Import",
        "tags": [
          "imports"
        ]
      }
    },
    "/v1/imports/{operation_id}": {
      "get": {
        "operationId": "read_import_v1_imports__operation_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "operation_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Operation Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OperationResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "401": {
            "description": "Unauthorized"
          },
          "404": {
            "description": "Not Found"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          },
          "503": {
            "description": "Service Unavailable"
          }
        },
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "summary": "Read Import",
        "tags": [
          "imports"
        ]
      }
    },
    "/v1/integrations/{operation_id}": {
      "get": {
        "operationId": "read_integration_v1_integrations__operation_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "operation_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Operation Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/IntegrationResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "401": {
            "description": "Unauthorized"
          },
          "404": {
            "description": "Not Found"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          },
          "503": {
            "description": "Service Unavailable"
          }
        },
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "summary": "Read Integration",
        "tags": [
          "integrations"
        ]
      }
    }
  }
}